fix: wrap /invite page in Suspense and clean up duplicate exports
